backend_fastapi_test/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.20
databases[postgresql]==0.6.0
asyncpg==0.27.0
pydantic==2.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.6.0
